Implement padding for BSQ outputs
Basic functionality works but it is not tested much and we discovered
more edge cases need to be handled. The handling of change outputs,
mixed BSQ/BTC outputs, missing handling of padding at other BSQ txs
like comp. request and voting made that feature overly complex and we
decided to leave it out for now. We are aware that adding it later
probably require a hard fork but I think the costs and risks are just
too high to add it now.
The simplicity of using plain BTC outputs without additional metadata as
core element would get lost by that feature.